+++ go-server/main.go
@@ -1,32 +1,21 @@
 package main
 
 import (
-	"context"
-	"database/sql"
-	"encoding/json"
 	"fmt"
 	"log"
 	"math/rand"
 	"net/http"
 	"sync"
-	"time"
 
 	"github.com/gorilla/websocket"
-	_ "github.com/lib/pq"
-	"github.com/redis/go-redis/v9"
 )
 
-type Server struct {
-	db                *sql.DB
-	redis             *redis.Client
-	clients           map[*websocket.Conn]*Client
-	rooms             map[string]map[*websocket.Conn]*Client
-	roomSubscriptions map[string]*redis.PubSub
-	mutex             sync.RWMutex
-	otEngine          *OTEngine
-	spatialIndex      *SpatialIndex
-	compressionManager *CompressionManager
-	sessionRecovery   *SessionRecovery
+// Simple server without Redis, OT, or spatial indexing
+type SimpleServer struct {
+	clients map[*websocket.Conn]*Client
+	rooms   map[string]map[*websocket.Conn]*Client
+	strokes map[string][]Stroke // room -> strokes (in-memory for now)
+	mutex   sync.RWMutex
 }
 
 type Client struct {
@@ -43,21 +32,22 @@ type Message struct {
 }
 
 type Stroke struct {
-	Points   [][]float64 `json:"points"`
-	Color    string      `json:"color"`
-	Thickness float64    `json:"thickness"`
-	IsEraser bool        `json:"isEraser"`
+	ID        string      `json:"id,omitempty"`
+	Points    [][]float64 `json:"points"`
+	Color     string      `json:"color"`
+	Thickness float64     `json:"thickness"`
+	IsEraser  bool        `json:"isEraser"`
 }
 
 var upgrader = websocket.Upgrader{
 	CheckOrigin: func(r *http.Request) bool {
-		return true // Allow all origins for hackathon
+		return true
 	},
 }
 
 func generateUsername() string {
-	adjectives := []string{"Swift", "Bright", "Clever", "Quick", "Smart", "Fast", "Sharp", "Wise", "Bold", "Calm"}
-	nouns := []string{"Penguin", "Dragon", "Phoenix", "Tiger", "Eagle", "Wolf", "Lion", "Bear", "Fox", "Hawk"}
+	adjectives := []string{"Swift", "Bright", "Clever", "Quick", "Smart"}
+	nouns := []string{"Penguin", "Dragon", "Phoenix", "Tiger", "Eagle"}
 	
 	adj := adjectives[rand.Intn(len(adjectives))]
 	noun := nouns[rand.Intn(len(nouns))]
@@ -66,7 +56,7 @@ func generateUsername() string {
 	return fmt.Sprintf("%s%s%d", adj, noun, num)
 }
 
-func (s *Server) handleWebSocket(w http.ResponseWriter, r *http.Request) {
+func (s *SimpleServer) handleWebSocket(w http.ResponseWriter, r *http.Request) {
 	conn, err := upgrader.Upgrade(w, r, nil)
 	if err != nil {
 		log.Printf("WebSocket upgrade failed: %v", err)
@@ -74,9 +64,7 @@ func (s *Server) handleWebSocket(w http.ResponseWriter, r *http.Request) {
 	}
 	defer conn.Close()
 
-	client := &Client{
-		conn: conn,
-	}
+	client := &Client{conn: conn}
 	
 	s.mutex.Lock()
 	s.clients[conn] = client
@@ -85,17 +73,11 @@ func (s *Server) handleWebSocket(w http.ResponseWriter, r *http.Request) {
 	defer func() {
 		s.mutex.Lock()
 		delete(s.clients, conn)
-		// Remove client from room
 		if client.room != "" {
 			if roomClients, exists := s.rooms[client.room]; exists {
 				delete(roomClients, conn)
-				// If room is empty, close Redis subscription
 				if len(roomClients) == 0 {
-					if pubsub := s.roomSubscriptions[client.room]; pubsub != nil {
-						pubsub.Close()
-						delete(s.roomSubscriptions, client.room)
-						delete(s.rooms, client.room)
-					}
+					delete(s.rooms, client.room)
 				}
 			}
 		}
@@ -116,42 +98,30 @@ func (s *Server) handleWebSocket(w http.ResponseWriter, r *http.Request) {
 			s.handleJoin(client, msg)
 		case "stroke":
 			s.handleStroke(client, msg)
-		case "operation":
-			s.handleOperation(client, msg)
 		case "clear":
 			s.handleClear(client, msg)
 		case "chat:message":
-			s.handleChatMessage(client, msg)
+			s.handleChat(client, msg)
 		}
 	}
 }
 
-func (s *Server) handleJoin(client *Client, msg Message) {
+func (s *SimpleServer) handleJoin(client *Client, msg Message) {
 	client.room = msg.Room
 	client.username = generateUsername()
 
 	s.mutex.Lock()
-	// Initialize room if it doesn't exist
 	if s.rooms[client.room] == nil {
 		s.rooms[client.room] = make(map[*websocket.Conn]*Client)
 	}
 	s.rooms[client.room][client.conn] = client
-
-	// Start Redis subscription for room if not already subscribed
-	if s.roomSubscriptions[client.room] == nil {
-		pubsub := s.redis.Subscribe(context.Background(), "room:"+client.room)
-		s.roomSubscriptions[client.room] = pubsub
-		go s.subscribeToRoom(client.room)
-	}
 	s.mutex.Unlock()
 
-	// Load existing strokes for the room
-	strokes, err := s.loadRoomStrokes(client.room)
-	if err != nil {
-		log.Printf("Error loading strokes: %v", err)
-	}
+	// Send existing strokes to new client
+	s.mutex.RLock()
+	strokes := s.strokes[client.room]
+	s.mutex.RUnlock()
 
-	// Send join confirmation
 	response := Message{
 		Type:     "joined",
 		Room:     client.room,
@@ -160,510 +130,204 @@ func (s *Server) handleJoin(client *Client, msg Message) {
 	}
 	
 	client.conn.WriteJSON(response)
-
 	log.Printf("Client %s joined room %s", client.username, client.room)
 }
 
-func (s *Server) handleStroke(client *Client, msg Message) {
-	// Resolve room ID by name for OT/persistence consistency
-	roomID, err := s.getRoomIDByName(client.room)
-	if err != nil {
-		log.Printf("Failed to resolve room ID for room %s: %v", client.room, err)
+func (s *SimpleServer) handleStroke(client *Client, msg Message) {
+	log.Printf("Received stroke message from %s: %+v", client.username, msg)
+	
+	// Extract stroke data
+	dataMap, ok := msg.Data.(map[string]interface{})
+	if !ok {
+		log.Printf("Invalid stroke data format - expected map[string]interface{}, got %T", msg.Data)
+		log.Printf("Raw data: %+v", msg.Data)
 		return
 	}
 
-	// Create operation for OT processing
-	operation := &Operation{
-		Type:   OpStrokeCreate,
-		RoomID: roomID,
-		UserID: client.username,
-		Data:   map[string]interface{}{
-			"stroke_data": msg.Data,
-		},
-	}
+	log.Printf("Stroke data map: %+v", dataMap)
 
-	// Add client version if provided
-	if clientVersion, ok := msg.Data.(map[string]interface{})["client_version"]; ok {
-		operation.Data["client_version"] = clientVersion
-	}
+	// Create stroke with timestamp for last-write-wins
+	stroke := Stroke{}
 
-	// Process operation through OT engine
-	result, err := s.otEngine.ProcessOperation(operation)
-	if err != nil {
-		log.Printf("Error processing stroke operation: %v", err)
-		// Send error back to client
-		errorMsg := Message{
-			Type: "error",
-			Room: client.room,
-			Data: map[string]interface{}{
-				"message": "Failed to process stroke",
-				"error":   err.Error(),
-			},
-		}
-		client.conn.WriteJSON(errorMsg)
-		return
-	}
-
-	if !result.Success {
-		log.Printf("Operation transformation failed: %v", result.Error)
-		return
-	}
-
-	// Save stroke with spatial data
-	err = s.saveStrokeWithSpatial(result.Operation)
-	if err != nil {
-		log.Printf("Error saving stroke: %v", err)
-		return
-	}
-
-	// Broadcast transformed operation to all clients
-	broadcastMsg := Message{
-		Type:     "operation",
-		Room:     client.room,
-		Username: result.Operation.UserID,
-		Data: map[string]interface{}{
-			"operation": result.Operation,
-		},
+	// Extract stroke properties
+	if id, ok := dataMap["id"].(string); ok {
+		stroke.ID = id
+		log.Printf("Extracted stroke ID: %s", id)
 	}
 	
-	broadcastJSON, _ := json.Marshal(broadcastMsg)
-	s.redis.Publish(context.Background(), "room:"+client.room, broadcastJSON)
-	// Local broadcast to connected clients to avoid subscription timing gaps
+	if pointsInterface, ok := dataMap["points"].([]interface{}); ok {
+		// Convert []interface{} to [][]float64
+		points := make([][]float64, len(pointsInterface))
+		for i, pointInterface := range pointsInterface {
+			if pointArray, ok := pointInterface.([]interface{}); ok {
+				point := make([]float64, len(pointArray))
+				for j, coord := range pointArray {
+					if coordFloat, ok := coord.(float64); ok {
+						point[j] = coordFloat
+					} else {
+						log.Printf("Invalid coordinate type at [%d][%d]: %T", i, j, coord)
+					}
+				}
+				points[i] = point
+			} else {
+				log.Printf("Invalid point type at [%d]: %T", i, pointInterface)
+			}
+		}
+		stroke.Points = points
+		log.Printf("Extracted points: %+v", points)
+	} else {
+		log.Printf("Failed to extract points, got type: %T", dataMap["points"])
+		log.Printf("Points value: %+v", dataMap["points"])
+	}
+	
+	if color, ok := dataMap["color"].(string); ok {
+		stroke.Color = color
+		log.Printf("Extracted color: %s", color)
+	} else {
+		log.Printf("Failed to extract color, got type: %T", dataMap["color"])
+		log.Printf("Color value: %+v", dataMap["color"])
+	}
+	
+	if thickness, ok := dataMap["thickness"].(float64); ok {
+		stroke.Thickness = thickness
+		log.Printf("Extracted thickness: %f", thickness)
+	} else if thicknessInt, ok := dataMap["thickness"].(int); ok {
+		stroke.Thickness = float64(thicknessInt)
+		log.Printf("Extracted thickness (converted from int): %f", stroke.Thickness)
+	} else {
+		log.Printf("Failed to extract thickness, got type: %T", dataMap["thickness"])
+		log.Printf("Thickness value: %+v", dataMap["thickness"])
+	}
+	
+	if isEraser, ok := dataMap["isEraser"].(bool); ok {
+		stroke.IsEraser = isEraser
+		log.Printf("Extracted isEraser: %t", isEraser)
+	} else {
+		log.Printf("Failed to extract isEraser, got type: %T", dataMap["isEraser"])
+		log.Printf("IsEraser value: %+v", dataMap["isEraser"])
+	}
+
+	log.Printf("Final stroke to be stored: %+v", stroke)
+	
+	// Add or update stroke in room
+	s.mutex.Lock()
+	if s.strokes[client.room] == nil {
+		s.strokes[client.room] = []Stroke{}
+	}
+	
+	// If stroke has an ID, try to update existing stroke
+	if stroke.ID != "" {
+		found := false
+		for i, existingStroke := range s.strokes[client.room] {
+			if existingStroke.ID == stroke.ID {
+				s.strokes[client.room][i] = stroke
+				found = true
+				log.Printf("Updated existing stroke with ID: %s", stroke.ID)
+				break
+			}
+		}
+		if !found {
+			s.strokes[client.room] = append(s.strokes[client.room], stroke)
+			log.Printf("Added new stroke with ID: %s", stroke.ID)
+		}
+	} else {
+		// No ID, add as new stroke
+		s.strokes[client.room] = append(s.strokes[client.room], stroke)
+		log.Printf("Added new stroke without ID")
+	}
+	s.mutex.Unlock()
+
+	// Broadcast to all clients in room
+	broadcastMsg := Message{
+		Type:     "stroke",
+		Room:     client.room,
+		Username: client.username,
+		Data:     stroke,
+	}
+
+	log.Printf("Broadcasting stroke message: %+v", broadcastMsg)
+
 	s.mutex.RLock()
 	if roomClients := s.rooms[client.room]; roomClients != nil {
 		for conn := range roomClients {
-			_ = conn.WriteMessage(websocket.TextMessage, broadcastJSON)
+			if conn != client.conn { // Don't send back to sender
+				conn.WriteJSON(broadcastMsg)
+			}
 		}
 	}
 	s.mutex.RUnlock()
+
+	log.Printf("Stroke from %s broadcast to room %s", client.username, client.room)
 }
 
-// handleOperation processes full Operation payloads from clients (Redux path)
-func (s *Server) handleOperation(client *Client, msg Message) {
-    // Resolve room ID
-    roomID, err := s.getRoomIDByName(client.room)
-    if err != nil {
-        log.Printf("Failed to resolve room ID for room %s: %v", client.room, err)
-        return
-    }
+func (s *SimpleServer) handleClear(client *Client, msg Message) {
+	s.mutex.Lock()
+	s.strokes[client.room] = []Stroke{}
+	s.mutex.Unlock()
 
-    // msg.Data should be the operation object
-    dataMap, ok := msg.Data.(map[string]interface{})
-    if !ok {
-        log.Printf("operation data is not an object: %T", msg.Data)
-        return
-    }
-
-    // Build Operation
-    op := &Operation{
-        Type:   OpStrokeCreate,
-        RoomID: roomID,
-        UserID: client.username,
-        Data:   map[string]interface{}{},
-    }
-
-    if t, ok := dataMap["type"].(string); ok {
-        op.Type = t
-    }
-    if uid, ok := dataMap["userId"].(string); ok && uid != "" {
-        op.UserID = uid
-    }
-    if d, ok := dataMap["data"].(map[string]interface{}); ok {
-        // Expect stroke_data nested for stroke_create
-        if sd, ok := d["stroke_data"]; ok {
-            op.Data["stroke_data"] = sd
-        } else {
-            // If client sent raw stroke fields, wrap them
-            op.Data["stroke_data"] = d
-        }
-        if cv, ok := d["client_version"]; ok {
-            op.Data["client_version"] = cv
-        }
-    } else {
-        // If entire dataMap is stroke_data
-        op.Data["stroke_data"] = dataMap
-    }
-
-    // Process via OT
-    result, err := s.otEngine.ProcessOperation(op)
-    if err != nil || !result.Success {
-        if err != nil {
-            log.Printf("Error processing operation: %v", err)
-        } else {
-            log.Printf("Operation transformation failed: %v", result.Error)
-        }
-        errorMsg := Message{
-            Type: "error",
-            Room: client.room,
-            Data: map[string]interface{}{
-                "message": "Failed to process operation",
-                "error":   result.Error,
-            },
-        }
-        client.conn.WriteJSON(errorMsg)
-        return
-    }
-
-    // Persist stroke with spatial index if stroke create
-    if result.Operation.Type == OpStrokeCreate {
-        if err := s.saveStrokeWithSpatial(result.Operation); err != nil {
-            log.Printf("Error saving stroke: %v", err)
-            return
-        }
-    }
-
-    // Broadcast
-    broadcastMsg := Message{
-        Type:     "operation",
-        Room:     client.room,
-        Username: result.Operation.UserID,
-        Data: map[string]interface{}{
-            "operation": result.Operation,
-        },
-    }
-    broadcastJSON, _ := json.Marshal(broadcastMsg)
-    s.redis.Publish(context.Background(), "room:"+client.room, broadcastJSON)
-    s.mutex.RLock()
-    if roomClients := s.rooms[client.room]; roomClients != nil {
-        for conn := range roomClients {
-            _ = conn.WriteMessage(websocket.TextMessage, broadcastJSON)
-        }
-    }
-    s.mutex.RUnlock()
-}
-
-func (s *Server) handleClear(client *Client, msg Message) {
-	// Clear strokes from PostgreSQL
-	err := s.clearRoomStrokes(client.room)
-	if err != nil {
-		log.Printf("Error clearing strokes: %v", err)
-		return
-	}
-
-	// Broadcast clear message
-	clearMsg := Message{
+	// Broadcast clear to all clients
+	broadcastMsg := Message{
 		Type:     "clear",
 		Room:     client.room,
 		Username: client.username,
+		Data:     nil,
 	}
-	
-	clearJSON, _ := json.Marshal(clearMsg)
-	s.redis.Publish(context.Background(), "room:"+client.room, clearJSON)
-	// Local broadcast
+
 	s.mutex.RLock()
 	if roomClients := s.rooms[client.room]; roomClients != nil {
 		for conn := range roomClients {
-			_ = conn.WriteMessage(websocket.TextMessage, clearJSON)
-		}
-	}
-	s.mutex.RUnlock()
-}
-
-func (s *Server) loadRoomStrokes(room string) ([]Stroke, error) {
-	rows, err := s.db.Query(`
-		SELECT stroke_data FROM strokes 
-		WHERE room_id = (SELECT id FROM rooms WHERE name = $1)
-		ORDER BY created_at ASC
-	`, room)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var strokes []Stroke
-	for rows.Next() {
-		var strokeData []byte
-		if err := rows.Scan(&strokeData); err != nil {
-			return nil, err
-		}
-		
-		var stroke Stroke
-		if err := json.Unmarshal(strokeData, &stroke); err != nil {
-			return nil, err
-		}
-		strokes = append(strokes, stroke)
-	}
-	
-	return strokes, nil
-}
-
-// saveStrokeWithSpatial saves stroke with spatial bounding box for indexing
-func (s *Server) saveStrokeWithSpatial(op *Operation) error {
-	strokeData, ok := op.Data["stroke_data"]
-	if !ok {
-		return fmt.Errorf("no stroke_data in operation")
-	}
-
-	strokeJSON, err := json.Marshal(strokeData)
-	if err != nil {
-		return err
-	}
-
-	// Calculate bounding box from stroke points
-	bbox, err := s.calculateStrokeBoundingBox(strokeData)
-	if err != nil {
-		return fmt.Errorf("failed to calculate bounding box: %v", err)
-	}
-
-	// Insert stroke with spatial data and OT version (use room_id directly)
-	// Map OT operation type to strokes.operation_type enum
-	opType := "create"
-	switch op.Type {
-	case OpStrokeCreate:
-		opType = "create"
-	case OpStrokeUpdate:
-		opType = "update"
-	case OpStrokeDelete:
-		opType = "delete"
-	}
-
-	_, err = s.db.Exec(`
-		INSERT INTO strokes (id, room_id, user_id, stroke_data, version, operation_type, bbox, created_at)
-		VALUES (
-			$1,
-			$2,
-			$3, $4, $5, $6,
-			box(point($7,$8), point($9,$10)),
-			$11
-		)`,
-		op.ID, op.RoomID, op.UserID, strokeJSON, op.Version, opType,
-		bbox.X1, bbox.Y1, bbox.X2, bbox.Y2, op.CreatedAt,
-	)
-	
-	if err != nil {
-		return err
-	}
-
-	// Add to spatial index for efficient viewport queries
-	indexedStroke := &IndexedStroke{
-		ID:       op.ID,
-		RoomID:   op.RoomID,
-		UserID:   op.UserID,
-		Data:     strokeData,
-		Version:  op.Version,
-		BBox:     *bbox,
-		IsActive: true,
-	}
-
-	return s.spatialIndex.Insert(indexedStroke)
-}
-
-// calculateStrokeBoundingBox calculates spatial bounding box for stroke
-func (s *Server) calculateStrokeBoundingBox(strokeData interface{}) (*BoundingBox, error) {
-	// Default bounding box
-	bbox := &BoundingBox{X1: 0, Y1: 0, X2: 100, Y2: 100}
-
-	// Parse stroke data to extract points
-	strokeMap, ok := strokeData.(map[string]interface{})
-	if !ok {
-		return bbox, nil
-	}
-
-	points, ok := strokeMap["points"].([]interface{})
-	if !ok || len(points) == 0 {
-		return bbox, nil
-	}
-
-	// Initialize with first point
-	firstPoint, ok := points[0].([]interface{})
-	if !ok || len(firstPoint) < 2 {
-		return bbox, nil
-	}
-
-	x, ok1 := firstPoint[0].(float64)
-	y, ok2 := firstPoint[1].(float64)
-	if !ok1 || !ok2 {
-		return bbox, nil
-	}
-
-	bbox.X1, bbox.Y1 = x, y
-	bbox.X2, bbox.Y2 = x, y
-
-	// Expand bounding box for all points
-	for _, pointInterface := range points[1:] {
-		point, ok := pointInterface.([]interface{})
-		if !ok || len(point) < 2 {
-			continue
-		}
-
-		x, ok1 := point[0].(float64)
-		y, ok2 := point[1].(float64)
-		if !ok1 || !ok2 {
-			continue
-		}
-
-		if x < bbox.X1 {
-			bbox.X1 = x
-		}
-		if x > bbox.X2 {
-			bbox.X2 = x
-		}
-		if y < bbox.Y1 {
-			bbox.Y1 = y
-		}
-		if y > bbox.Y2 {
-			bbox.Y2 = y
-		}
-	}
-
-	// Add some padding for stroke thickness
-	padding := 10.0
-	bbox.X1 -= padding
-	bbox.Y1 -= padding
-	bbox.X2 += padding
-	bbox.Y2 += padding
-
-	return bbox, nil
-}
-
-func (s *Server) clearRoomStrokes(room string) error {
-	_, err := s.db.Exec(`
-		DELETE FROM strokes 
-		WHERE room_id = (SELECT id FROM rooms WHERE name = $1)
-	`, room)
-	
-	return err
-}
-
-func (s *Server) subscribeToRoom(room string) {
-	s.mutex.RLock()
-	pubsub := s.roomSubscriptions[room]
-	s.mutex.RUnlock()
-
-	if pubsub == nil {
-		log.Printf("No subscription found for room: %s", room)
-		return
-	}
-
-	defer func() {
-		s.mutex.Lock()
-		if s.roomSubscriptions[room] == pubsub {
-			pubsub.Close()
-			delete(s.roomSubscriptions, room)
-		}
-		s.mutex.Unlock()
-	}()
-
-	for {
-		msg, err := pubsub.ReceiveMessage(context.Background())
-		if err != nil {
-			log.Printf("Redis receive error for room %s: %v", room, err)
-			break
-		}
-
-		// Broadcast to all clients in the room
-		s.mutex.RLock()
-		roomClients := s.rooms[room]
-		if roomClients != nil {
-			log.Printf("Broadcasting message to %d clients in room %s", len(roomClients), room)
-			for conn := range roomClients {
-				err := conn.WriteMessage(websocket.TextMessage, []byte(msg.Payload))
-				if err != nil {
-					log.Printf("Failed to send message to client, removing connection: %v", err)
-					// Clean up dead connection
-					conn.Close()
-					s.mutex.RUnlock()
-					s.mutex.Lock()
-					if rc, ok := s.rooms[room]; ok {
-						delete(rc, conn)
-					}
-					s.mutex.Unlock()
-					s.mutex.RLock()
-				}
+			if conn != client.conn {
+				conn.WriteJSON(broadcastMsg)
 			}
 		}
-		s.mutex.RUnlock()
 	}
+	s.mutex.RUnlock()
+
+	log.Printf("Canvas cleared by %s in room %s", client.username, client.room)
 }
 
-// handleChatMessage broadcasts chat messages to all clients in the room via Redis
-func (s *Server) handleChatMessage(client *Client, msg Message) {
-	chatMsg := Message{
+func (s *SimpleServer) handleChat(client *Client, msg Message) {
+	log.Printf("Handling chat message from %s in room %s", client.username, client.room)
+
+	var chatData interface{}
+	if payload, ok := msg.Data.(map[string]interface{})["payload"]; ok {
+        chatData = payload
+    } else {
+        // Fallback to direct data
+        chatData = msg.Data
+    }
+    
+
+	broadcastMsg := Message{
 		Type:     "chat:message",
 		Room:     client.room,
 		Username: client.username,
-		Data:     msg.Data,
+		Data:     chatData,
 	}
-	chatJSON, _ := json.Marshal(chatMsg)
-	s.redis.Publish(context.Background(), "room:"+client.room, chatJSON)
-	// Local broadcast
+
 	s.mutex.RLock()
 	if roomClients := s.rooms[client.room]; roomClients != nil {
 		for conn := range roomClients {
-			_ = conn.WriteMessage(websocket.TextMessage, chatJSON)
+			if conn != client.conn {
+				conn.WriteJSON(broadcastMsg)
+			}
 		}
 	}
 	s.mutex.RUnlock()
 }
 
-// getRoomIDByName resolves a room name to its ID
-func (s *Server) getRoomIDByName(roomName string) (string, error) {
-	var id string
-	err := s.db.QueryRow("SELECT id FROM rooms WHERE name = $1", roomName).Scan(&id)
-	if err != nil {
-		return "", err
-	}
-	return id, nil
-}
-
 func main() {
-	// Note: Go 1.20+ automatically seeds random number generator
-
-	// Connect to PostgreSQL
-	db, err := sql.Open("postgres", "postgres://postgres:password@localhost:5432/whiteboard?sslmode=disable")
-	if err != nil {
-		log.Fatal("Failed to connect to PostgreSQL:", err)
+	server := &SimpleServer{
+		clients: make(map[*websocket.Conn]*Client),
+		rooms:   make(map[string]map[*websocket.Conn]*Client),
+		strokes: make(map[string][]Stroke),
 	}
-	defer db.Close()
 
-	// Test database connection
-	err = db.Ping()
-	if err != nil {
-		log.Fatal("Failed to ping PostgreSQL:", err)
-	}
-	log.Println("Connected to PostgreSQL")
-
-	// Connect to Redis
-	redisClient := redis.NewClient(&redis.Options{
-		Addr: "localhost:6379",
-		DB:   0,
+	http.HandleFunc("/ws", server.handleWebSocket)
+	http.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
+		w.Write([]byte("OK"))
 	})
 
-	// Test Redis connection
-	ctx := context.Background()
-	_, err = redisClient.Ping(ctx).Result()
-	if err != nil {
-		log.Fatal("Failed to connect to Redis:", err)
-	}
-	log.Println("Connected to Redis")
-
-	// Create server with OT engine and spatial index
-	server := &Server{
-		db:                db,
-		redis:             redisClient,
-		clients:           make(map[*websocket.Conn]*Client),
-		rooms:             make(map[string]map[*websocket.Conn]*Client),
-		roomSubscriptions: make(map[string]*redis.PubSub),
-	}
-	
-	// Initialize OT engine, spatial index, compression, and session recovery
-	server.otEngine = NewOTEngine(server)
-	server.spatialIndex = NewSpatialIndex()
-	server.compressionManager = NewCompressionManager(10, 100*time.Millisecond) // Batch 10 messages or 100ms timeout
-	server.sessionRecovery = NewSessionRecovery(server)
-	
-	// Start cleanup routines
-	server.sessionRecovery.StartCleanupRoutine()
-
-	// Setup routes
-	http.HandleFunc("/ws", server.handleWebSocket)
-	http.HandleFunc("/api/viewport", server.handleViewportQuery)
-	http.HandleFunc("/api/stats/spatial", server.handleSpatialStats)
-	http.HandleFunc("/health", server.handleHealthCheck)
-	
-	// Serve static files for testing
-	http.Handle("/", http.FileServer(http.Dir("../frontend/dist")))
-
-	log.Println("Server starting on :8080")
+	log.Println("Simple WebSocket server starting on :8080")
 	log.Fatal(http.ListenAndServe(":8080", nil))
 }
